VCL-26 Fixed "Subroutine get_reservation_remote_ip redefined" error messages.
Changed DataStructure.pm::get_reservation_remote_ip to return undefined if an error occurred, and 0 if the remote IP isn't set. This makes the return values consistent with the rest of the code.

Updated reserved.pm to use the subroutine in DataStructure.pm instead of the old one in utils.pm. Also updated it to recognize the new return values.

Removed old utils.pm::get_reservation_remote_ip subroutine.

git-svn-id: https://svn.apache.org/repos/asf/incubator/vcl/trunk@735090 13f79535-47bb-0310-9956-ffa450edef68