VCL-73 Added a sanity check to new.pm::computer_not_being_used if a neighbor process was in the maintenance state. If it finds this, it returns 0 immediately. This state will occur if an imaging reservation failed. The computer cannot be reloaded until the imaging problem is fixed.
If new.pm::computer_not_being_used attempted to kill a process but the exit status returned was not 0, it could hang if the process was actually killed. There is a bug with Perl 5.8.0 which causes exit statuses to be reported incorrectly. I removed the loop continuation if kill looks like it failed. This is safer than the possibility of in infinite loop because it will only get to that point if the reservation process being killed isn't another user's process.

git-svn-id: https://svn.apache.org/repos/asf/incubator/vcl/trunk@739350 13f79535-47bb-0310-9956-ffa450edef68