VCL-100
added a conditional check to account for new distro - This is the only location on the backend that still depends on the distro name. This will be removed at a future date once a method for generalizing Linux distros is pushed to the database or we get away from using distros in the prefix of the internal imagename. 



git-svn-id: https://svn.apache.org/repos/asf/incubator/vcl/trunk@751700 13f79535-47bb-0310-9956-ffa450edef68