VCL-117
Removed hardcoded value for the return-path sendmail -f option in the
utils:mail routine

created a global value that gets set in the vcld.conf file RETURNPATH

NOTE: the previous utils.pm was a premature check-in of work being done on healthcheck. this checkin are the changes
related to the return-path variable.



git-svn-id: https://svn.apache.org/repos/asf/incubator/vcl/trunk@757529 13f79535-47bb-0310-9956-ffa450edef68