VCL-123
This commit adds basic Ubuntu support through an Ubuntu.pm OS module using the new OS module framework corresponding to the VCL-123 ticket.  The Linux.pm isn't adequate to handle the user adding because Ubuntu doesn't support setting the password using --stdin which is what the core defaults to if there is not OS module to tell it how to do it otherwise.  Additionally, the core isn't suitable to handle the user adding because it contains the '-g ncsu' field which is not appropriate for places outside of NCSU.



git-svn-id: https://svn.apache.org/repos/asf/incubator/vcl/trunk@760101 13f79535-47bb-0310-9956-ffa450edef68