VCL-20 Added ability to DataStructure.pm to retrieve management node's ssh port. Extended DataStructure.pm to allow the management node to be specified when using any of the get_management_node_xxx() functions.
VCL-82
Modified regex's in utils.pm::notify() to remove blank lines from the log output and to remove spaces from the beginning of lines.

VCL-22
Modified utils.pm::get_management_node_info() to accept an IP address as the argument. It will return the info for the management node with the IP specified.

Other
Made minor change to utils.pm::run_ssh_command() output. It displays the full ssh command in the log if verbose mode is enabled. I also modified it to not display the return values if the 'no output' option was specified.

Added an optional 'no output' argument to utils.pm::run_command(). This was done to control the output and to match run_ssh_command().

Changed a few 'OK' notify() messages to 'DEBUG' so they wouldn't appear in the log file if verbose mode is disabled.

git-svn-id: https://svn.apache.org/repos/asf/incubator/vcl/trunk@774398 13f79535-47bb-0310-9956-ffa450edef68