VCL-135 Fixed a bug which was introduced in r774413. The xCAT template path was not being returned correctly if linux_image path was used.
Other
Changed xCAT.pm::node_status() so that it returns 'RELOAD' if the nodetype.tab file does not match what is in the currentimage.txt file on the node. It had been checking for this but ignoring mismatches, assuming that nodetype.tab was correct.

Fixed typos in xCAT.pm and xCAT21.pm where 'CRITCAL' should have been 'CRITICAL'.

git-svn-id: https://svn.apache.org/repos/asf/incubator/vcl/trunk@775946 13f79535-47bb-0310-9956-ffa450edef68