VCL-191
Added paths for vsphere 4:
module checks for the installed locations, based on the known paths it
sets it either to vsphere4 or the previous vmware toolkit

Added generic get_vmware_host_info routine to fetch various vmware
server information. For this issue it is used to collect the true
hostname of the esx server. In future additions it looks like it could  be used to
determine available capicity 

Added additional guestOS variables

Added option for eth0 and eth1 vmware generated. The
VMWARE_MAC_ETH0_GENERATED and VMWARE_MAC_ETH1_GENERATED
are currently set in the vcld.conf file. If these are not set it skips
the arp checks to find the vmware assigned MAC addresses 




git-svn-id: https://svn.apache.org/repos/asf/incubator/vcl/trunk@807191 13f79535-47bb-0310-9956-ffa450edef68